recorder/pyproject.toml
[project]
name = "recorder"
version = "0.1.0"
description = "Standalone teleop data stream monitor for G1"
requires-python = "==3.10.*"
dependencies = [
    "numpy",
    "pyzmq",
    "opencv-python",
    "msgpack",
    "ipdb",
    "elements",
    "portal",
    "unitree_sdk2py",
]

[tool.uv.sources]
unitree_sdk2py = { path = "../external_dependencies/unitree_sdk2_python", editable = true }

# unitree_sdk2py depends on cyclonedds==0.10.2, whose pip build needs
# CYCLONEDDS_HOME pointing at the matching C library install (e.g.
# /home/unitree/cyclonedds-0.10.2/install). Export it before `uv sync`.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["stream_monitor.py"]

[tool.hatch.build.targets.wheel.sources]
"stream_monitor.py" = "recorder/stream_monitor.py"
